scanners/iods/requirements/base.txt
# IODS Core Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
rich>=13.0.0
structlog>=23.0.0
pyyaml>=6.0.0
jinja2>=3.1.0
click>=8.1.0
python-multipart>=0.0.6
httpx>=0.25.0
aiofiles>=23.0.0
